First pass at including new driver
