+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "priam",
-  "version": "1.2.1",
+  "version": "2.0.0",
   "description": "A simple Cassandra driver. It wraps the helenus and cassandra-driver modules with additional error/retry handling, external .cql file support, connection option resolution from an external source, and query composition, among other improvements.",
   "keywords": [
     "cassandra",
