+++ test/baseline-cover.js
@@ -1,5 +1,4 @@
 require('../lib/driver');
-require('../lib/drivers/datastax');
 require('../lib/util/batch');
 require('../lib/util/query');
 require('../lib/util/query-cache');
